group user file transfer tests together; add xfailing tests for #215
