+++ tests/integration/user_file_transfer/test_automatic_file_transfer.py
